requirements.txt
flask
flask-login
flask-sqlalchemy
PyPDF2
python-docx
werkzeug
